k/quotas: quota_manager refactoring
ctor moved to cc
throttle_delay::enforce instead of first_violation to better express intent
