+++ src/v/kafka/server/quota_manager.cc
@@ -23,6 +23,27 @@ namespace kafka {
 using clock = quota_manager::clock;
 using throttle_delay = quota_manager::throttle_delay;
 
+quota_manager::quota_manager()
+  : _default_num_windows(config::shard_local_cfg().default_num_windows.bind())
+  , _default_window_width(config::shard_local_cfg().default_window_sec.bind())
+  , _default_target_produce_tp_rate(
+      config::shard_local_cfg().target_quota_byte_rate.bind())
+  , _default_target_fetch_tp_rate(
+      config::shard_local_cfg().target_fetch_quota_byte_rate.bind())
+  , _target_partition_mutation_quota(
+      config::shard_local_cfg().kafka_admin_topic_api_rate.bind())
+  , _target_produce_tp_rate_per_client_group(
+      config::shard_local_cfg().kafka_client_group_byte_rate_quota.bind())
+  , _target_fetch_tp_rate_per_client_group(
+      config::shard_local_cfg().kafka_client_group_fetch_byte_rate_quota.bind())
+  , _gc_freq(config::shard_local_cfg().quota_manager_gc_sec())
+  , _max_delay(config::shard_local_cfg().max_kafka_throttle_delay_ms.bind()) {
+    _gc_timer.set_callback([this] {
+        auto full_window = _default_num_windows() * _default_window_width();
+        gc(full_window);
+    });
+}
+
 quota_manager::~quota_manager() { _gc_timer.cancel(); }
 
 ss::future<> quota_manager::stop() {
@@ -227,7 +248,7 @@ throttle_delay quota_manager::record_produce_tp_and_throttle(
     auto prev = it->second.delay;
     it->second.delay = delay_ms;
     throttle_delay res{};
-    res.first_violation = prev.count() == 0;
+    res.enforce = prev.count() > 0;
     res.duration = it->second.delay;
     return res;
 }
@@ -256,7 +277,7 @@ throttle_delay quota_manager::throttle_fetch_tp(
     auto delay_ms = throttle(
       quota_id, *target_tp_rate, now, it->second.tp_fetch_rate);
     throttle_delay res{};
-    res.first_violation = false;
+    res.enforce = true;
     res.duration = delay_ms;
     return res;
 }
