+++ src/v/kafka/server/quota_manager.h
@@ -10,7 +10,8 @@
  */
 
 #pragma once
-#include "config/configuration.h"
+#include "config/client_group_byte_rate_quota.h"
+#include "config/property.h"
 #include "kafka/server/token_bucket_rate_tracker.h"
 #include "resource_mgmt/rate.h"
 #include "seastarx.h"
@@ -54,40 +55,22 @@ public:
     using clock = ss::lowres_clock;
 
     struct throttle_delay {
-        bool first_violation{true};
-        clock::duration duration{std::chrono::milliseconds(0)};
+        bool enforce{false};
+        clock::duration duration{0};
+        clock::duration enforce_duration() const {
+            if (enforce) {
+                return duration;
+            } else {
+                return clock::duration::zero();
+            }
+        }
     };
 
-    quota_manager()
-      : _default_num_windows(
-        config::shard_local_cfg().default_num_windows.bind())
-      , _default_window_width(
-          config::shard_local_cfg().default_window_sec.bind())
-      , _default_target_produce_tp_rate(
-          config::shard_local_cfg().target_quota_byte_rate.bind())
-      , _default_target_fetch_tp_rate(
-          config::shard_local_cfg().target_fetch_quota_byte_rate.bind())
-      , _target_partition_mutation_quota(
-          config::shard_local_cfg().kafka_admin_topic_api_rate.bind())
-      , _target_produce_tp_rate_per_client_group(
-          config::shard_local_cfg().kafka_client_group_byte_rate_quota.bind())
-      , _target_fetch_tp_rate_per_client_group(
-          config::shard_local_cfg()
-            .kafka_client_group_fetch_byte_rate_quota.bind())
-      , _gc_freq(config::shard_local_cfg().quota_manager_gc_sec())
-      , _max_delay(
-          config::shard_local_cfg().max_kafka_throttle_delay_ms.bind()) {
-        _gc_timer.set_callback([this] {
-            auto full_window = _default_num_windows() * _default_window_width();
-            gc(full_window);
-        });
-    }
-
+    quota_manager();
     quota_manager(const quota_manager&) = delete;
     quota_manager& operator=(const quota_manager&) = delete;
     quota_manager(quota_manager&&) = delete;
     quota_manager& operator=(quota_manager&&) = delete;
-
     ~quota_manager();
 
     ss::future<> stop();
